refactor: moved components folder to correct location
