+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 import Head from "next/head"
 
-import Navbar from "./components/navbar/navbar"
+import Navbar from "../components/navbar/navbar"
 
 export default function Home() {
   return (
@@ -10,9 +10,6 @@ export default function Home() {
         <title>Huhu Leirisovellus</title>
       </Head>
       <main className="relative bg-helsinki w-full h-full overflow-hidden flex flex-col items-start justify-start box-border text-left text-[28px] text-palevioletred font-poppins">
-         
-      <Navbar></Navbar>
-
       </main>
     </div>
   )
